requirements.txt
Django==4.0
django-jazzmin==2.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
